+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Add `Context::toInteractive()` method
 * Add `Castor\Event\ContextCreatedEvent` to allow updating the context after it is created
 * Add `run_phar()` function to run a phar file in all contexts
+* Add `Context::withVerboseArguments()` method to pass verbose arguments to the underlying process when needed
 
 ### Vendor